// src/app/models/apply/application.ts
export class Application {
    name?: string;
    surname?: string;
    dateOfBirth?: any;
    dateApplied?: any;
    sex?: string;
    passportNumber?: string;
    passportIssueDate?: any;
    passportExpiryDate?: any;
    nationality?: string;
    institution?: string;
    course?: string;
    studyYear?: any;
    duration?: any;
    email?: any;
    phone?: any;
    residentialAddress?: string;
    status?: string;
    approved?: string;
    applicationType?: string;
    registrationProofFile?: any;
    academicRecordFile?: any;
    passportFile?: any;
    residenceProofFile?: any;
    criminalRecordNamibiaFile?: any;
    criminalRecordOriginFile?: any;

}